Forgot to move some classes
